Changed application prop to yml and added comments

+++ src/main/java/com/cfg/petadoption/entity/Pet.java
@@ -3,17 +3,17 @@ package com.cfg.petadoption.entity;
 import jakarta.persistence.*;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
+// Represents a pet stored in the database.
 @Entity
 @Table(name = "pets")
 
 @Data
 @NoArgsConstructor
 public class Pet {
+    // Automatically generated primary key.
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-
+    private long id;
     private String name;
     private String species;
     private String breed;
